Added Field level validation and error messages.
Here's a list of constraints; @NotNull checks if a property is not null,
@AssertTrue checks if a Boolean value is true, there's also an
@AssertFalse, @Min and @Max checks the minimum and the maximum values of
a number, but it could also be a string representing a number, @Size is
the length of a string or the length of a collection, @Digits represents
a number with an integer and a fraction, and a @Pattern is a regular
expression, @NotEmpty means the property is not null or empty, and
@NotBlank contains no whitespaces, @Positive means with a positive
number, there's also a @Negative, and there's also a @PositiveOrZero and
a @NegativeOrZero, @Email checks if the email address is valid, @Past
checks if a date is in the past, there's also @PastOrPresent, and
@Future checks if a date is in the future, and there's also a
@FutureOrPresent. Over to the demo. We add a validation to the friend
class, a @NotBlank. In the controller we add the @Valid and
@ExceptionHandler on the MethodArgumentNotValidException. In IntelliJ we
don't need the validation in the controller anymore. It moves to the
model to the friend. The only thing that's left here is an annotation
@Valid that will do the job. In the Friend class we add an annotation to
the firstName, @NotBlank. This is all.